ai/digimad_backend_ai_io_controller/requirements.txt
pydantic==1.7.2
fastapi==0.61.2
files==1.1.1
mysql_connector_repackaged==0.3.1
python-dotenv==0.15.0
mysql-connector-python==8.0.22
fastapi-jwt-auth==0.5.0
colorutils==0.3.0
requests==2.22.0
